+++ changelog.md
@@ -1,2 +1,2 @@
 ### Formations 1.0.1
-- Initial release of Formations for Forge on 1.19.4
+- Initial release of Formations for Forge on 1.19.2